feat(menu): handle menu image upload via menu module
